+++ src/v1/constants/errors.ts
@@ -33,15 +33,3 @@ export const ERRORS = {
     }
 }
 
-  
-export class LogicError extends Error {
-    code: number;
-    msg: string;
-    data: any;
-    constructor({ code, msg, data }: any, ...params: any) {
-      super(...params);
-      this.code = code;
-      this.msg = msg;
-      this.data = data;
-    }
-  }
